+++ tests/PboStudio.Tests/TestPlanBehaviourTests.cs
@@ -134,6 +134,9 @@ public class TestPlanBehaviourTests
         }
 
         public void Pause(TimeSpan duration) { }
+        public void Suspend() { }
+        public void Resume() { }
+
         public void Stop() { }
         public void Dispose() { }
     }